refactor: split redirects into individual files
