+++ apps/arclight/src/app/[...route]/route.ts
@@ -92,212 +92,22 @@ const setCorsHeaders = (c: any) => {
   c.header('Access-Control-Expose-Headers', '*')
 }
 
-const hlsRoute = createRoute({
-  method: 'get',
-  path: '/hls/:mediaComponentId/:languageId',
-  tags: ['Redirects by-convention'],
-  summary: 'Redirects to the HTTP Live Streaming (HLS) rendition',
-  description:
-    'Redirects to the HTTP Live Streaming (HLS) rendition of the given media component and language IDs.',
-  request: {
-    params: z.object({
-      mediaComponentId: z.string().describe('The ID of the media component'),
-      languageId: z.string().describe('The ID of the language')
-    })
-  },
-  responses: {
-    200: {
-      content: {
-        'application/json': {
-          schema: z.object({
-            url: z.string().describe('The HLS video URL')
-          })
-        }
-      },
-      description: 'Returns the HLS video URL'
-    },
-    302: {
-      description:
-        'Redirects to the HLS stream URL. Note: The handler currently returns 200 with JSON, this documents intended redirect behavior for HLS manifest/playlist.'
-    },
-    404: {
-      content: {
-        'application/json': {
-          schema: z.object({
-            error: z.string()
-          })
-        }
-      },
-      description: 'Video variant or HLS URL not found'
-    },
-    500: {
-      content: {
-        'application/json': {
-          schema: z.object({
-            error: z.string()
-          })
-        }
-      },
-      description: 'Internal server error'
-    }
+// Helper to get the client IP from Hono context
+function getClientIp(c: any): string | undefined {
+  const forwardedFor = c.req.header('x-forwarded-for')
+  if (forwardedFor) {
+    return forwardedFor.split(',')[0].trim()
   }
-} as const)
+  const realIp = c.req.header('x-real-ip')
+  if (realIp) {
+    return realIp
+  }
+  // Fallback for direct connections (like localhost)
+  return c.env?.remoteAddr?.address
+}
 
-const lowQualityRoute = createRoute({
-  method: 'get',
-  path: '/dl/:mediaComponentId/:languageId',
-  tags: ['Redirects by-convention'],
-  summary: 'Redirects to the smallest/lowest quality download rendition',
-  description:
-    'Redirects to the smallest/lowest quality download rendition of the given media component and language IDs.',
-  request: {
-    params: z.object({
-      mediaComponentId: z.string().describe('The ID of the media component'),
-      languageId: z.string().describe('The ID of the language')
-    })
-  },
-  responses: {
-    302: {
-      description: 'Redirects to the download URL'
-    },
-    404: {
-      content: {
-        'application/json': {
-          schema: z.object({
-            error: z.string()
-          })
-        }
-      },
-      description: 'Download URL not found'
-    },
-    500: {
-      content: {
-        'application/json': {
-          schema: z.object({
-            error: z.string()
-          })
-        }
-      },
-      description: 'Internal server error'
-    }
-  }
-} as const)
-
-const highQualityRoute = createRoute({
-  method: 'get',
-  path: '/dh/:mediaComponentId/:languageId',
-  tags: ['Redirects by-convention'],
-  summary: 'Redirects to the largest/highest quality download rendition',
-  description:
-    'Redirects to the largest/highest quality download rendition of the given media component and language IDs.',
-  request: {
-    params: z.object({
-      mediaComponentId: z.string().describe('The ID of the media component'),
-      languageId: z.string().describe('The ID of the language')
-    })
-  },
-  responses: {
-    302: {
-      description: 'Redirects to the high-quality download URL'
-    },
-    404: {
-      content: {
-        'application/json': {
-          schema: z.object({
-            error: z.string()
-          })
-        }
-      },
-      description: 'High-quality download URL not found'
-    },
-    500: {
-      content: {
-        'application/json': {
-          schema: z.object({
-            error: z.string()
-          })
-        }
-      },
-      description: 'Internal server error'
-    }
-  }
-} as const)
-
-const watchRoute = createRoute({
-  method: 'get',
-  path: '/s/:mediaComponentId/:languageId',
-  tags: ['Redirects by-convention'],
-  summary: 'Redirects to the target share URL',
-  description:
-    'Redirects to the target share URL of the given media component and language IDs.',
-  request: {
-    params: z.object({
-      mediaComponentId: z.string().describe('The ID of the media component'),
-      languageId: z.string().describe('The ID of the language')
-    })
-  },
-  responses: {
-    302: {
-      description: 'Redirects to the watch page'
-    },
-    500: {
-      content: {
-        'application/json': {
-          schema: z.object({
-            error: z.string()
-          })
-        }
-      },
-      description: 'Internal server error'
-    }
-  }
-} as const)
-
-const keywordRoute = createRoute({
-  method: 'get',
-  path: '/:keyword',
-  tags: ['Redirect keyword'],
-  summary: 'Redirects a short URL keyword',
-  description: 'Returns a short URL keyword to its corresponding long URL.',
-  request: {
-    params: z.object({
-      keyword: z
-        .string()
-        .regex(
-          /^[A-Za-z0-9_-]+$/,
-          'Keyword must contain only letters, numbers, dashes or underscores'
-        )
-        .min(3, 'Keyword must be at least 3 characters')
-        .max(32, 'Keyword cannot exceed 32 characters')
-        .describe('The short URL keyword')
-    })
-  },
-  responses: {
-    302: {
-      description: 'Redirects to the long URL'
-    },
-    404: {
-      content: {
-        'application/json': {
-          schema: z.object({
-            error: z.string()
-          })
-        }
-      },
-      description: 'Keyword not found'
-    },
-    500: {
-      content: {
-        'application/json': {
-          schema: z.object({
-            error: z.string()
-          })
-        }
-      },
-      description: 'Internal server error'
-    }
-  }
-} as const)
+// Remove all redirect route logic for hls, dl, dh, s, and keyword
+// Only keep non-redirect logic (e.g., OpenAPI docs, unrelated routes)
 
 app.doc('/redirects-doc.json', {
   openapi: '3.0.0',
@@ -309,212 +119,14 @@ app.doc('/redirects-doc.json', {
 
 app.get('/api/redirects-doc', swaggerUI({ url: '/redirects-doc.json' }))
 
-app.openapi(hlsRoute, async (c) => {
-  setCorsHeaders(c)
-  const { mediaComponentId, languageId } = c.req.param()
-  console.log('HLS request', { mediaComponentId, languageId })
-  try {
-    const variant = await getVideoVariant(mediaComponentId, languageId)
-    console.log('video variant', variant)
-
-    const brightcoveId = variant.brightcoveId ?? null
-    if (brightcoveId) {
-      console.log('Attempt Brightcove first', brightcoveId)
-      try {
-        const video = await getBrightcoveVideo(brightcoveId)
-        const src = selectBrightcoveSource(video, 'hls')
-        if (src) {
-          console.log('Brightcove src', src)
-          return c.redirect(src, 302)
-        } else {
-          console.warn('No Brightcove HLS source found for', brightcoveId)
-        }
-      } catch (err) {
-        console.error('Brightcove HLS fetch failed:', err)
-      }
-    }
-
-    // Fallback to variant.hls if Brightcove is not available or fails
-    if (variant.hls) {
-      console.log('Using variant.hls fallback', variant.hls)
-      // Optionally, you can fetch the URL to follow redirects, but here we just redirect
-      return c.redirect(variant.hls, 302)
-    }
-
-    return c.json({ error: 'HLS URL not available' }, 404)
-  } catch (error) {
-    if (error instanceof HTTPException) {
-      if (error.status === 404) {
-        return c.json({ error: error.message }, 404)
-      }
-    }
-    const errorMessage = error instanceof Error ? error.message : String(error)
-    return c.json({ error: `Internal server error: ${errorMessage}` }, 500)
-  }
-})
-
-app.openapi(lowQualityRoute, async (c) => {
-  setCorsHeaders(c)
-  const { mediaComponentId, languageId } = c.req.param()
-  log('DL request', { mediaComponentId, languageId })
-  try {
-    const variant = await getVideoVariant(mediaComponentId, languageId)
-    const download = variant.downloads?.find((d) => d.quality === 'low')
-
-    if (!download?.url) {
-      // Fallback to Brightcove
-      const brightcoveId = variant.brightcoveId ?? null
-      log('DL Brightcove fallback', brightcoveId)
-      if (brightcoveId) {
-        try {
-          const video = await getBrightcoveVideo(brightcoveId)
-          const src = selectBrightcoveSource(video, 'dl')
-          if (src) {
-            return c.redirect(src, 302)
-          }
-        } catch (err) {
-          console.error(
-            'Brightcove fallback for low quality download failed:',
-            err
-          )
-        }
-      }
-      return c.json({ error: 'Low quality download URL not available' }, 404)
-    }
-
-    return c.redirect(download.url)
-  } catch (error) {
-    if (error instanceof HTTPException) {
-      if (error.status === 404) {
-        return c.json({ error: error.message }, 404)
-      }
-    }
-    const errorMessage = error instanceof Error ? error.message : String(error)
-    return c.json({ error: `Internal server error: ${errorMessage}` }, 500)
-  }
-})
-
-app.openapi(highQualityRoute, async (c) => {
-  setCorsHeaders(c)
-  const { mediaComponentId, languageId } = c.req.param()
-  log('DH request', { mediaComponentId, languageId })
-  try {
-    const variant = await getVideoVariant(mediaComponentId, languageId)
-    const download = variant.downloads?.find((d) => d.quality === 'high')
-
-    if (!download?.url) {
-      // Fallback to Brightcove
-      const brightcoveId = variant.brightcoveId ?? null
-      log('DH Brightcove fallback', brightcoveId)
-      if (brightcoveId) {
-        try {
-          const video = await getBrightcoveVideo(brightcoveId)
-          const src = selectBrightcoveSource(video, 'dh')
-          if (src) {
-            return c.redirect(src, 302)
-          }
-        } catch (err) {
-          console.error(
-            'Brightcove fallback for high quality download failed:',
-            err
-          )
-        }
-      }
-      return c.json({ error: 'High quality download URL not available' }, 404)
-    }
-
-    return c.redirect(download.url)
-  } catch (error) {
-    if (error instanceof HTTPException) {
-      if (error.status === 404) {
-        return c.json({ error: error.message }, 404)
-      }
-    }
-    const errorMessage = error instanceof Error ? error.message : String(error)
-    return c.json({ error: `Internal server error: ${errorMessage}` }, 500)
-  }
-})
-
-app.openapi(watchRoute, async (c) => {
-  setCorsHeaders(c)
-  const { mediaComponentId, languageId } = c.req.param()
-  log('Share request', { mediaComponentId, languageId })
-  try {
-    const variant = await getVideoVariant(mediaComponentId, languageId)
-
-    if (variant.share) {
-      return c.redirect(variant.share, 302)
-    }
-
-    const brightcoveId = variant.brightcoveId ?? null
-    log('Share Brightcove fallback', brightcoveId)
-    if (brightcoveId) {
-      try {
-        const video = await getBrightcoveVideo(brightcoveId)
-        const src = selectBrightcoveSource(video, 's')
-        if (src) return c.redirect(src, 302)
-      } catch (err) {
-        console.error('Brightcove fallback for share URL failed:', err)
-      }
-    }
-
-    // Fallback to legacy watch page if all else fails
-    return c.redirect(
-      `https://jesusfilm.org/bin/jf/watch.html/${mediaComponentId}/${languageId}`,
-      302
-    )
-  } catch (error) {
-    const errorMessage = error instanceof Error ? error.message : String(error)
-    return c.json({ error: `Internal server error: ${errorMessage}` }, 500)
-  }
-})
-
-// --- Keyword short-link redirect -------------------------------------------------
-
-app.openapi(keywordRoute, async (c) => {
-  setCorsHeaders(c)
-  const { keyword } = c.req.param()
-
-  try {
-    const { data, error, errors } = await getApolloClient().query<
-      ResultOf<typeof GET_SHORT_LINK_QUERY>
-    >({
-      query: GET_SHORT_LINK_QUERY,
-      variables: {
-        hostname: 'arc.gt',
-        pathname: keyword
-      }
-    })
-
-    if (error != null || errors != null) {
-      console.error('GraphQL error fetching short link:', error ?? errors)
-      throw new HTTPException(500, {
-        message: 'Failed to query short link data'
-      })
-    }
-
-    if (
-      data?.shortLink?.__typename === 'QueryShortLinkByPathSuccess' &&
-      data.shortLink.data.to
-    ) {
-      return c.redirect(data.shortLink.data.to, 302)
-    }
-
-    return c.json({ error: 'Keyword not found or invalid' }, 404)
-  } catch (err) {
-    if (err instanceof HTTPException) {
-      throw err // Let Hono handle
-    }
-    console.error('Error resolving keyword redirect:', err)
-    const errorMessage = err instanceof Error ? err.message : String(err)
-    return c.json({ error: `Internal server error: ${errorMessage}` }, 500)
-  }
-})
-
-// -------------------------------------------------------------------------------
+// ... keep any other non-redirect routes here ...
 
 app.options('*', (c) => {
-  setCorsHeaders(c)
+  // Optionally keep CORS preflight for non-redirects
+  c.header('Access-Control-Allow-Origin', '*')
+  c.header('Access-Control-Allow-Methods', 'GET, OPTIONS')
+  c.header('Access-Control-Allow-Headers', '*')
+  c.header('Access-Control-Expose-Headers', '*')
   return c.body(null, 204)
 })
